renamed to chain_generic_tests
