+++ darkside-tests/src/chain_generic_tests.rs
@@ -1,18 +1,10 @@
-
-
-
-
-
 use zingolib::{
-    lightclient::LightClient, test_framework::generic_chain_tests::ChainTest,
-    wallet::WalletBase,
+    lightclient::LightClient, test_framework::chain_generic_tests::ChainTest, wallet::WalletBase,
 };
 
 use crate::{
     constants::{ABANDON_TO_DARKSIDE_SAP_10_000_000_ZAT, DARKSIDE_SEED},
-    utils::{
-        scenarios::DarksideScenario, update_tree_states_for_transaction,
-    },
+    utils::{scenarios::DarksideScenario, update_tree_states_for_transaction},
 };
 
 impl ChainTest for DarksideScenario {
@@ -88,7 +80,7 @@ impl ChainTest for DarksideScenario {
 
 #[tokio::test]
 async fn chain_generic_send() {
-    zingolib::test_framework::generic_chain_tests::simple_send::<DarksideScenario>(40_000).await;
+    zingolib::test_framework::chain_generic_tests::simple_send::<DarksideScenario>(40_000).await;
 }
 
 use proptest::proptest;
@@ -97,7 +89,7 @@ proptest! {
     #[test]
     fn chain_generic_send_proptest(value in 0..90_000u32) {
         Runtime::new().unwrap().block_on(async {
-    zingolib::test_framework::generic_chain_tests::simple_send::<DarksideScenario>(value).await;
+    zingolib::test_framework::chain_generic_tests::simple_send::<DarksideScenario>(value).await;
         });
      }
 }
