+++ darkside-tests/src/lib.rs
@@ -1,5 +1,5 @@
+pub mod chain_generic_tests;
 pub mod constants;
-pub mod generic_chain_tests;
 pub mod utils;
 pub mod darkside_types {
     tonic::include_proto!("cash.z.wallet.sdk.rpc");
