+++ zingolib/src/test_framework.rs
@@ -2,4 +2,4 @@
 pub(crate) mod macros;
 pub mod mocks;
 
-pub mod generic_chain_tests;
+pub mod chain_generic_tests;
